requirements.txt
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2019.3
whitenoise==5.3.0
pyTelegramBotAPI==4.2.2
django-heroku
ratelimiter
schedule==1.1.0
geopy==2.2.0
apscheduler==3.6.3
requests
bs4
tzlocal
python-dateutil
Pillow == 8.2.0
tronapi
pycoingecko
